refactor(playwright): decouple playwright deployment code into init-deployment generator
