+++ packages/playwright/generators.json
@@ -8,6 +8,11 @@
             "schema": "./src/generators/init/schema.json",
             "description": "init generator"
         },
+        "init-deployment": {
+            "factory": "./src/generators/init-deployment/generator",
+            "schema": "./src/generators/init-deployment/schema.json",
+            "description": "The deployment generator contains all the pipeline and infrastructure change for playwright."
+        },
         "accessibility": {
             "factory": "./src/generators/accessibility/generator",
             "schema": "./src/generators/accessibility/schema.json",
