+++ packages/playwright/src/generators/init-deployment/schema.d.ts
@@ -1,0 +1,2 @@
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
+export interface InitDeploymentGeneratorSchema {}
