--- packages/playwright/src/generators/init/utils/versions.ts
@@ -1,1 +1,0 @@
-export const PLAYWRIGHT_VERSION = '^1.31.1';
