+++ packages/playwright/src/index.ts
@@ -1,1 +1,2 @@
 export { default as initGenerator } from './generators/init/generator';
+export { default as initDeploymentGenerator } from './generators/init-deployment/generator';
