+++ packages/playwright/src/utils/versions.ts
@@ -1,3 +1,4 @@
 export const AXE_CORE_PLAYWRIGHT_VERSION = '4.5.2';
 export const AXE_RESULTS_PRETTY_PRINT_VERSION = '1.0.2';
 export const APPLITOOLS_EYES_PLAYWRIGHT_VERSION = '1.13.0';
+export const PLAYWRIGHT_VERSION = '^1.31.1';
